Move form for map to seperate component
